refactor: add ImportResultsPageClient & ImportResultsTable
