+++ apps/data-service-catalog/app/catalogs/[catalogId]/data-services/import-results/import-results-page-client.module.css
@@ -1,3 +1,15 @@
+.select {
+  width: fit-content;
+}
+
+.chips {
+  margin-top: 20px;
+}
+
+.wrap {
+  flex-wrap: wrap;
+}
+
 .importFilter {
   display: flex;
   padding: 1px;
