+++ libs/ui/src/index.ts
@@ -26,7 +26,7 @@ export * from './lib/header';
 export * from './lib/help-markdown';
 export * from './lib/icon';
 export * from './lib/import-result-details';
-export * from './lib/import-results';
+export * from './lib/import-results-table';
 export * from './lib/info-card';
 export * from './lib/intersection-observer';
 export * from './lib/key-value-list';
